+++ src/dev/start_dev_server.js
@@ -6,16 +6,23 @@ import {
 import { Abort, raceProcessTeardownEvents } from "@jsenv/abort"
 import { createLogger, createTaskLog } from "@jsenv/log"
 import { getCallerPosition } from "@jsenv/urls"
+import {
+  jsenvAccessControlAllowedHeaders,
+  startServer,
+  jsenvServiceCORS,
+  jsenvServiceErrorHandler,
+} from "@jsenv/server"
+import { convertFileSystemErrorToResponseProperties } from "@jsenv/server/src/internal/convertFileSystemErrorToResponseProperties.js"
 
+import { createServerEventsDispatcher } from "@jsenv/core/src/plugins/server_events/server_events_dispatcher.js"
 import { defaultRuntimeCompat } from "@jsenv/core/src/build/build.js"
 import { createReloadableWorker } from "@jsenv/core/src/helpers/worker_reload.js"
-import { startOmegaServer } from "@jsenv/core/src/omega/omega_server.js"
+import { createFileService } from "./file_service.js"
 
 export const startDevServer = async ({
   signal = new AbortController().signal,
   handleSIGINT = true,
   logLevel = "info",
-  omegaServerLogLevel = "warn",
   protocol = "http",
   // it's better to use http1 by default because it allows to get statusText in devtools
   // which gives valuable information when there is errors
@@ -26,7 +33,8 @@ export const startDevServer = async ({
   port = 3456,
   acceptAnyIp,
   keepProcessAlive = true,
-  services,
+  services = [],
+  onStop = () => {},
 
   rootDirectoryUrl,
   clientFiles = {
@@ -134,39 +142,130 @@ export const startDevServer = async ({
     disabled: !logger.levels.info,
   })
 
-  const server = await startOmegaServer({
-    logLevel: omegaServerLogLevel,
+  const serverStopCallbacks = []
+  const serverEventsDispatcher = createServerEventsDispatcher()
+  serverStopCallbacks.push(() => {
+    serverEventsDispatcher.destroy()
+  })
+  const server = await startServer({
+    signal,
+    stopOnExit: false,
+    stopOnSIGINT: handleSIGINT,
+    stopOnInternalError: false,
     keepProcessAlive,
-    acceptAnyIp,
+    logLevel,
+    startLog: false,
+
     protocol,
     http2,
     certificate,
     privateKey,
+    acceptAnyIp,
     hostname,
     port,
-    services,
+    requestWaitingMs: 60_000,
+    services: [
+      jsenvServiceCORS({
+        accessControlAllowRequestOrigin: true,
+        accessControlAllowRequestMethod: true,
+        accessControlAllowRequestHeaders: true,
+        accessControlAllowedRequestHeaders: [
+          ...jsenvAccessControlAllowedHeaders,
+          "x-jsenv-execution-id",
+        ],
+        accessControlAllowCredentials: true,
+        timingAllowOrigin: true,
+      }),
+      ...services,
+      {
+        name: "jsenv:omega_file_service",
+        handleRequest: createFileService({
+          signal,
+          logLevel,
+          serverStopCallbacks,
+          serverEventsDispatcher,
 
-    rootDirectoryUrl,
-    scenarios: { dev: true },
-    runtimeCompat,
+          rootDirectoryUrl,
+          scenarios: { dev: true },
+          runtimeCompat,
 
-    plugins,
-    urlAnalysis,
-    urlResolution,
-    fileSystemMagicRedirection,
-    supervisor,
-    transpilation,
-    clientFiles,
-    clientMainFileUrl,
-    clientAutoreload,
-    cooldownBetweenFileEvents,
-    explorer,
-    sourcemaps,
-    sourcemapsSourcesProtocol,
-    sourcemapsSourcesContent,
-    writeGeneratedFiles,
+          plugins,
+          urlAnalysis,
+          urlResolution,
+          fileSystemMagicRedirection,
+          supervisor,
+          transpilation,
+          clientAutoreload,
+          clientFiles,
+          clientMainFileUrl,
+          cooldownBetweenFileEvents,
+          explorer,
+          sourcemaps,
+          sourcemapsSourcesProtocol,
+          sourcemapsSourcesContent,
+          writeGeneratedFiles,
+        }),
+        handleWebsocket: (websocket, { request }) => {
+          if (request.headers["sec-websocket-protocol"] === "jsenv") {
+            serverEventsDispatcher.addWebsocket(websocket, request)
+          }
+        },
+      },
+      {
+        name: "jsenv:omega_error_handler",
+        handleError: (error) => {
+          const getResponseForError = () => {
+            if (error && error.asResponse) {
+              return error.asResponse()
+            }
+            if (
+              error &&
+              error.statusText === "Unexpected directory operation"
+            ) {
+              return {
+                status: 403,
+              }
+            }
+            return convertFileSystemErrorToResponseProperties(error)
+          }
+          const response = getResponseForError()
+          if (!response) {
+            return null
+          }
+          const body = JSON.stringify({
+            status: response.status,
+            statusText: response.statusText,
+            headers: response.headers,
+            body: response.body,
+          })
+          return {
+            status: 200,
+            headers: {
+              "content-type": "application/json",
+              "content-length": Buffer.byteLength(body),
+            },
+            body,
+          }
+        },
+      },
+      // default error handling
+      jsenvServiceErrorHandler({
+        sendErrorDetails: true,
+      }),
+    ],
+    onStop: (reason) => {
+      onStop()
+      serverStopCallbacks.forEach((serverStopCallback) => {
+        serverStopCallback(reason)
+      })
+      serverStopCallbacks.length = 0
+    },
   })
   startDevServerTask.done()
+  if (hostname) {
+    delete server.origins.localip
+    delete server.origins.externalip
+  }
   logger.info(``)
   Object.keys(server.origins).forEach((key) => {
     logger.info(`- ${server.origins[key]}`)