+++ src/plugins/transpilation/as_js_classic/jsenv_plugin_as_js_classic_library.js
@@ -1,4 +1,4 @@
-import { createUrlGraphLoader } from "@jsenv/core/src/omega/url_graph/url_graph_loader.js"
+import { createUrlGraphLoader } from "@jsenv/core/src/kitchen/url_graph/url_graph_loader.js"
 import { bundleJsModules } from "@jsenv/core/src/plugins/bundling/js_module/bundle_js_modules.js"
 
 import { convertJsModuleToJsClassic } from "./convert_js_module_to_js_classic.js"