+++ src/plugins/url_resolution/node_esm_resolver.js
@@ -26,11 +26,17 @@ export const createNodeEsmResolver = ({ runtimeCompat, packageConditions }) => {
   ]
 
   return (reference, context) => {
-    const { parentUrl, specifier } = reference
+    if (reference.type === "package_json") {
+      return reference.specifier
+    }
+    const parentUrl = reference.baseUrl || reference.parentUrl
+    if (!parentUrl.startsWith("file:")) {
+      return new URL(reference.specifier, parentUrl).href
+    }
     const { url, type, packageUrl } = applyNodeEsmResolution({
       conditions: packageConditions,
       parentUrl,
-      specifier,
+      specifier: reference.specifier,
     })
     if (context.scenarios.dev) {
       const dependsOnPackageJson =